backend/requirements.txt
fastapi==0.127.1
uvicorn==0.22.0
pydantic==2.12.5

# Authentication
google-auth==2.45.0
google-auth-httplib2==0.3.0
Authlib==1.6.6
PyJWT==2.10.1
PyPDF2>=3.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
python-multipart>=0.0.9
alembic==1.11.1


# Database
psycopg2-binary==2.9.11
SQLAlchemy==2.0.45
python-dotenv==1.2.1
itsdangerous
itsdangerous==2.2.0


# Google API
google-api-python-client==2.187.0
google-api-core==2.28.1
google-genai==1.56.0

# LLM / embeddings
numpy==1.26.4
langchain==1.1.0
chromadb==0.4.22
chroma-hnswlib==0.7.3
torch==2.9.1
transformers==4.57.3
sentence-transformers==2.3.1

# Scraping / data
beautifulsoup4==4.12.3
requests==2.32.5
lxml==6.0.2
html5lib==1.1
